+++ frontend/src/app/login/page.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import KakaoBtn from '@/app/_components/ui/button/Kakao';
 
 import { useSession } from 'next-auth/react';
-import LoginLogo from './_components/ui/loginLogo/LoginLogo';
+import LoginLogo from './_components/ui/LoginLogo';
 
 export default function LoginPage() {
   return (
